+++ maiconverter/maisxt/maisxt.py
@@ -9,7 +9,7 @@ from .sxtnote import (
     HoldNote,
     SlideStartNote,
     SlideEndNote,
-    is_slide_valid,
+    check_slide,
 )
 from ..event import NoteType
 from ..tool import measure_to_second, second_to_measure
@@ -319,8 +319,7 @@ class MaiSxt:
         duration: float,
         pattern: int,
         delay: float = 0.25,
-        check_slide: bool = True,
-        fail_on_undefined: bool = False,
+        slide_check: bool = True,
         decrement: bool = True,
     ) -> None:
         """Adds both a start slide and end slide note to the list of notes.
@@ -335,9 +334,7 @@ class MaiSxt:
             pattern: Numerical representation of the slide pattern.
             delay: Duration from when the slide appears and when it
                 starts to move, in terms of measures. Defaults to 0.25.
-            check_slide: When set to true, will check validity of slides.
-            fail_on_undefined: Optional bool defaults to False. When True, slides with undefined
-                behaviour will raise a ValueError instead.
+            slide_check: When set to true, will check validity of slides.
             decrement: When set to true, measure is subtracted by 1. Defaults to true.
 
         Raises:
@@ -352,10 +349,8 @@ class MaiSxt:
             >>> sxt = MaiSxt()
             >>> sxt.add_slide(2, 6, 3, 1.75, 1)
         """
-        if check_slide and not is_slide_valid(
-            pattern, start_position, end_position, fail_on_undefined
-        ):
-            print(f"Warning: Adding slide with undefined behaviour.")
+        if slide_check:
+            check_slide(pattern, start_position, end_position)
 
         if decrement:
             measure = max(0.0, measure - 1.0)